+++ Reflection/src/utopia/reflection/container/stack/template/scrolling/ScrollViewLike2.scala
@@ -23,12 +23,12 @@ trait ScrollViewLike2[C <: Stackable2] extends ScrollAreaLike2[C]
 	  * @return The length of this scroll view
 	  */
 	def length = lengthAlong(axis)
-	def length_=(newLength: Double) = setLength(newLength, axis)
+	def length_=(newLength: Double) = setLength(axis(newLength))
 	/**
 	  * @return The breadth of this scroll view
 	  */
 	def breadth = lengthAlong(axis.perpendicular)
-	def breadth_=(newBreadth: Double) = setLength(newBreadth, axis.perpendicular)
+	def breadth_=(newBreadth: Double) = setLength(axis.perpendicular(newBreadth))
 	
 	/**
 	  * @return The length of this view's contents